Reverted (mostly) blen_waveforms.py
I was headed down the wrong path for developing the filter/shutter indicators. Hugo Slepicka (slepicka@) and Matt Gibbs (mgibbs@) guided me toward enum buttons as opposed to bespoke classes and signals/slots.

I added a ref to a PV ending in :RWF for the raw plots - this will likely change again in the future.