Another rework of the Attenuation widgets.
I got inspiration from the line edit widgets in cpswTreeGui. The line
edit fires off a monitor thread that watches the RBV for the Attenuation
register. If the widget is not being edited and the value in the
register is different from the value in the widget the text gets updated
to match the register. I had to go totally outside of the PyDMLineEdit
this time and subclass QLineEdit directly since the PyDMLineEdit does
not support this kind of behavior. I also used pyepics PVs to connect to
the CA endpoints.